refactor : PlayerService 분리
